Add inner firmware magic
I guess this is here because only firmware.code gets loaded into PMEM,
which does not include body.magic, and they want to be able to check the
firmware type when the chip is running. But then why even have
body.magic in the first place? Why not just check firmware.magic? Maybe
they have tools that operate on the firmware at different levels, and
some of those tools don't know how to look inside firmware.code to get
version/type information?

It is a mystery.