Add serial monitor firmware
Currently only works on the ASM1142, since I don't know the base address
of the UART on any of the other chips. Memory writes are also not
working fully--only 16 bits are written instead of the full 32 bits.
This code is also missing support for single-byte writes, which are
important because many of the MMIO registers only occupy one byte. I
also still don't fully understand the UART, like which bits of the TX
Status Register represent "FIFO/TX register empty" vs. "UART is finished
transmitting", or what many of the other registers do.