Build images for all targets simultaneously
It was hard to keep track of what chip I had built monitor.img for at
any given time, so it's probably best to just build for all targets at
the same time. In addition, split the image generation process so that
both flash images and binaries that can be written directly to CODE RAM
are produced.