Document array/vector math hardware
This isn't entirely correct, since there appears to be some meaning to
the lowest three bits of MISC_CTRL, but I haven't been able to get the
hardware to work when bit 0 is set, and bits 1 and 2 didn't seem to
change any of the results. Also, I'm still not completely certain
whether I want to call this "array" math hardware of "vector" math
hardware.