Add asm_tool.py
This is meant to be an easy-to-use tool for interacting with ASMedia USB
host controllers over the PCIe interface. Functionality will include
reading and writing MMIO registers via the hardware interface (for the
chips that support it), loading firmware via the hardware interface (for
the chips that support it), and more.