Automatically unbind the kernel driver when necessary to mmap BAR0
When the driver is bound to the device, the kernel prevents mmapping the
device's resources (though it seems that adding "iomem=relaxed" to the
kernel command line may disable this protection). To make sure that we
can always access the BAR0 registers, always attempt to unbind the
driver before performing the mmap.