Add loosely-organized notes on developting an xHC emulator
This is mostly just a brain dump of the features, goals, requirements,
challenges, etc. I thought up while brainstorming the architecture for a
new 8051 emulator. The point was to make the architecture flexible
enough to avoid a complete redesign every time I wanted to add a new
feature, and to enable it to be a base for future emulators for other
(non-xHC) 8051-based microcontrollers.